+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Changelog
 
 - Implemented UI for native token features [PR 2292](https://github.com/input-output-hk/daedalus/pull/2292)
 
+### Chores
+
+- Disabled application menu navigation before the "Terms of use" have been accepted ([PR 2304](https://github.com/input-output-hk/daedalus/pull/2304))
+
 ## 3.2.0-FC1
 
 ### Features